StoryReaderAPI/requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
oauthlib==3.1.0
Pillow==7.0.0
python-decouple==3.3
requests==2.22.0
urllib3==1.25.8

